Move ConfigValue out of mcpconfig, clean up
