+++ src/main/java/net/fabricmc/loom/configuration/providers/forge/mcpconfig/steplogic/InjectLogic.java
@@ -32,7 +32,7 @@ import java.nio.file.StandardCopyOption;
 import java.util.Iterator;
 import java.util.stream.Stream;
 
-import net.fabricmc.loom.configuration.providers.forge.mcpconfig.ConfigValue;
+import net.fabricmc.loom.configuration.providers.forge.ConfigValue;
 import net.fabricmc.loom.util.FileSystemUtil;
 
 public final class InjectLogic implements StepLogic {
